refactor: add base layout for all endpoints
